clustermesh: Consolidate hive
Several of the core Cells that are reused by clustermesh and kvstoremesh
are declared separately. In order to simplify the logic, this commit
moves those common Cells into a single declaration at the top level and
directly passes that common Cell into each Hive when it is created on
startup.

Signed-off-by: Joe Stringer <joe@cilium.io>